refactor(application): extract PipelineExecutor from PipelineBase
Extract state machine logic into dedicated PipelineExecutor class:
- Move _RunState dataclass and run phase methods to executor.py
- PipelineBase.run() now delegates to PipelineExecutor.execute()
- Executor returns RunResult(success=False) instead of raising on error
- Update tests to check result.success instead of catching exceptions
